[Demangle] Add support for D programming language
This patch adds support for D programming language demangling on LLVM core by
porting libiberty D demangler to LLVM C++ codebase. This will allow easier
integration on a future LLDB plugin for D either in the upstream tree or
outside of it. It also includes tests and a fuzzer tool.

Signed-off-by: Luís Ferreira <contact@lsferreira.net>